+++ baybe/__init__.py
@@ -2,10 +2,11 @@
 
 import warnings
 
-from baybe.campaign import Campaign
-
 # Show deprecation warnings
 warnings.filterwarnings("default", category=DeprecationWarning, module="baybe")
+
+from baybe.campaign import Campaign
+from baybe.settings import Settings, active_settings
 
 
 def infer_version() -> str:  # pragma: no cover
@@ -55,6 +56,8 @@ __version__ = infer_version()
 __all__ = [
     "__version__",
     "Campaign",
+    "Settings",
+    "active_settings",
 ]
 
 del infer_version