+++ baybe/campaign.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import gc
 import json
+import warnings
 from collections.abc import Callable, Collection, Sequence
 from functools import reduce
 from typing import TYPE_CHECKING, Any, TypeVar
@@ -21,6 +22,7 @@ from baybe.exceptions import (
     IncompatibilityError,
     NoMeasurementsError,
     NotEnoughPointsLeftError,
+    NothingToComputeError,
 )
 from baybe.objectives.base import Objective, to_objective
 from baybe.parameters.base import Parameter
@@ -37,21 +39,16 @@ from baybe.searchspace.core import (
     validate_searchspace_from_config,
 )
 from baybe.serialization import SerialMixin, converter
+from baybe.settings import Settings, active_settings
 from baybe.surrogates.base import PosteriorStatistic, SurrogateProtocol
 from baybe.targets.base import Target
 from baybe.utils.basic import UNSPECIFIED, UnspecifiedType, is_all_instance
 from baybe.utils.boolean import eq_dataframe
 from baybe.utils.conversion import to_string
-from baybe.utils.dataframe import (
-    _ValidatedDataFrame,
-    filter_df,
-    fuzzy_row_match,
-    normalize_input_dtypes,
-)
+from baybe.utils.dataframe import filter_df, fuzzy_row_match
 from baybe.utils.validation import (
+    preprocess_dataframe,
     validate_object_names,
-    validate_objective_input,
-    validate_parameter_input,
     validate_target_input,
 )
 
@@ -318,15 +315,13 @@ class Campaign(SerialMixin):
             numerical_measurements_must_be_within_tolerance: Flag indicating if
                 numerical parameters need to be within their tolerances.
         """
-        # Validate target and parameter input values
-        validate_target_input(data, self.targets)
-        if self.objective is not None:
-            validate_objective_input(data, self.objective)
-        validate_parameter_input(
-            data, self.parameters, numerical_measurements_must_be_within_tolerance
+        # Preprocess incoming data
+        data = preprocess_dataframe(
+            data,
+            self.searchspace,
+            self.objective,
+            numerical_measurements_must_be_within_tolerance,
         )
-        data = normalize_input_dtypes(data, self.parameters + self.targets)
-        data.__class__ = _ValidatedDataFrame
 
         # With new measurements, the recommendations must always be recomputed
         self.clear_cache()
@@ -369,15 +364,13 @@ class Campaign(SerialMixin):
             ValueError: If the given data contains indices not present in existing
                 measurements.
         """
-        # Validate target and parameter input values
-        validate_target_input(data, self.targets)
-        if self.objective is not None:
-            validate_objective_input(data, self.objective)
-        validate_parameter_input(
-            data, self.parameters, numerical_measurements_must_be_within_tolerance
+        # Preprocess incoming data
+        data = preprocess_dataframe(
+            data,
+            self.searchspace,
+            self.objective,
+            numerical_measurements_must_be_within_tolerance,
         )
-        data = normalize_input_dtypes(data, self.parameters + self.targets)
-        data.__class__ = _ValidatedDataFrame
 
         # With changed measurements, the recommendations must always be recomputed
         self.clear_cache()
@@ -500,19 +493,21 @@ class Campaign(SerialMixin):
         # IMPROVE: Currently, we simply invalidate the cache whenever pending
         #     experiments are provided, because in order to use it, we need to check if
         #     the previous call was done with the same pending experiments.
-
         if pending_experiments is not None:
             self.clear_cache()
 
-            validate_parameter_input(pending_experiments, self.parameters)
-            pending_experiments = normalize_input_dtypes(
-                pending_experiments, self.parameters
+        # Preprocess pending experiments
+        if pending_experiments is not None:
+            pending_experiments = preprocess_dataframe(
+                pending_experiments,
+                self.searchspace,
+                numerical_measurements_must_be_within_tolerance=False,
             )
-            pending_experiments.__class__ = _ValidatedDataFrame
 
         if (
-            pending_experiments is None
+            active_settings.cache_campaign_recommendations
             and (cache := self._cached_recommendation) is not None
+            and pending_experiments is None
             and self.allow_recommending_already_recommended
             and len(cache) == batch_size
         ):
@@ -567,15 +562,16 @@ class Campaign(SerialMixin):
 
         # Get the recommended search space entries
         try:
-            # NOTE: The `recommend` call must happen on `self.recommender` to update
-            #   potential inner states in case of meta recommenders!
-            rec = self.recommender.recommend(
-                batch_size,
-                searchspace,
-                self.objective,
-                self._measurements_exp,
-                None if is_nonpredictive else pending_experiments,
-            )
+            with Settings(preprocess_dataframes=False):
+                # NOTE: The `recommend` call must happen on `self.recommender` to update
+                #   potential inner states in case of meta recommenders!
+                rec = self.recommender.recommend(
+                    batch_size,
+                    searchspace,
+                    self.objective,
+                    self._measurements_exp,
+                    None if is_nonpredictive else pending_experiments,
+                )
         except NotEnoughPointsLeftError as ex:
             # Aliases for code compactness
             f = fields(Campaign)
@@ -606,7 +602,10 @@ class Campaign(SerialMixin):
                 f"{str(ex)} Consider setting {message}."
             ) from ex
 
-        if pending_experiments is None:  # see IMPROVE comment above
+        if (
+            active_settings.cache_campaign_recommendations
+            and pending_experiments is None  # see IMPROVE comment above
+        ):
             self._cache_recommendation(rec)
 
         # Update metadata
@@ -871,6 +870,70 @@ class Campaign(SerialMixin):
             pending_experiments,
             acquisition_function,
         )
+
+    def identify_non_dominated_configurations(
+        self,
+        configurations: pd.DataFrame | None = None,
+        /,
+        *,
+        consider_campaign_measurements: bool = True,
+    ) -> pd.Series:
+        """Create a Boolean mask indicating non-dominated configurations.
+
+        Args:
+            configurations: A dataframe carrying values for all targets tracked by the
+                campaign's objective. If ``None``, uses the campaign's measurements.
+            consider_campaign_measurements: If ``True``, the campaign's measurements are
+                considered for identifying the non-dominated configurations, but will
+                not be reflected in the returned mask themselves.
+
+        Raises:
+            IncompatibilityError: If no objective is defined for the campaign.
+            NothingToComputeError: If no configurations are provided as argument and no
+                measurements are added to the campaign yet.
+
+        Returns:
+            A Boolean series indicating which configurations are non-dominated.
+        """
+        if self.objective is None:
+            raise IncompatibilityError(
+                "Cannot get the non-dominated configurations since no "
+                f"'{Objective.__name__}' is defined."
+            )
+
+        if self.measurements.empty:
+            if configurations is None:
+                raise NothingToComputeError(
+                    "The calculation of non-dominated points was requested, but "
+                    "neither are configurations provided nor does the campaign have "
+                    "any measurements added yet. Therefore, there is nothing to "
+                    "compute."
+                )
+            if consider_campaign_measurements:
+                warnings.warn(
+                    "No measurements have been added to the campaign yet, but the flag "
+                    "`consider_campaign_measurements` is set to "
+                    "'True'. Therefore, the non-dominated configurations will be "
+                    "determined without taking any measurements into account.",
+                    UserWarning,
+                )
+
+        if configurations is None:
+            configurations = self.measurements
+        else:
+            validate_target_input(configurations, self.objective.targets)
+
+        if consider_campaign_measurements and not self.measurements.empty:
+            configurations = pd.concat([configurations, self.measurements])
+
+        non_dominated = self.objective.identify_non_dominated_configurations(
+            configurations
+        )
+
+        if consider_campaign_measurements:
+            non_dominated = non_dominated.iloc[: -len(self.measurements)]
+
+        return non_dominated
 
 
 def _add_version(dict_: dict) -> dict: