+++ benchmarks/domains/direct_arylation/convergence_tl.py
@@ -16,9 +16,9 @@ from baybe.parameters import (
 )
 from baybe.parameters.base import DiscreteParameter
 from baybe.searchspace import SearchSpace
+from baybe.settings import Settings
 from baybe.simulation import simulate_scenarios
 from baybe.targets import NumericalTarget
-from baybe.utils.random import temporary_seed
 from benchmarks.data.utils import DATA_PATH
 from benchmarks.definition import (
     ConvergenceBenchmarkSettings,
@@ -137,7 +137,7 @@ def direct_arylation_tl_temperature(
     percentages = [0.01, 0.1, 0.2]
 
     initial_data_samples = {}
-    with temporary_seed(settings.random_seed):
+    with Settings(random_seed=settings.random_seed):
         for p in percentages:
             initial_data_samples[p] = [
                 initial_data.sample(frac=p) for _ in range(settings.n_mc_iterations)